[SPARK-11505][SQL] Break aggregate functions into multiple files
